modify product category 2 level response in product catgory creating api

+++ src/main/java/ecommerce/example/ecommerce/models/Product.java
@@ -32,7 +32,7 @@ public class Product {
     private String description;
 
     @Column(name = "rating")
-    private Integer rating;
+    private Float rating;
 
     @Column(name = "total_sold")
     private Integer totalSold;
@@ -82,7 +82,7 @@ public class Product {
     public void prePersist() {
 
         this.totalSold = 0;
-        this.rating = 0;
+        this.rating = 0f;
         this.createdAt = LocalDateTime.now();
         this.updatedAt = LocalDateTime.now();
     }
